Added toggle-switch.jsx file.
